+++ SRCPDCCSlave/src/dcc/GAOpenDCC.h
@@ -19,7 +19,7 @@ private:
 	int startAddr;
 	int endAddr;
 public:
-	GAOpenDCC( int startAddr, uint8_t endAddr, srcp::SRCPGenericAccessoire* next );
+	GAOpenDCC( int startAddr, int endAddr, srcp::SRCPGenericAccessoire* next );
 	int set( int addr, int port, int value, int delay );
 	int checkAddr( int addr ) { return ( addr >= startAddr && addr <= endAddr); }
 };
